Change StaticImage to GatsbyImage
The thumbnail was previously a static image that causes issues with
smaller devices with resizing. This change reworks these images to
become dynamic GatsbyImage, this enables the dynamic resizing.
Further improvements also include the use of the `gatsby-source-filesystem`
plugin that enables the retrieval of the images with `graphql`, this
retrieves the image and then further enhances it.